create a 10x10 nparray

+++ py/ray/scipy_ray/ndriver.py
@@ -3,11 +3,11 @@ from numpy import loadtxt
 
 
 def read_array(fn: str) -> np.array:
-    arr = loadtxt(fn, comments="#", delimiter=",", unpack=False)
-    return arr
+    arr = loadtxt(fn, comments="#", delimiter=',', unpack=False)
+    return arr.astype('int')
 
 
-def add_array(a1: np.array, a2:np.array) -> np.array:
+def add_array(a1: np.array, a2: np.array) -> np.array:
     return np.add(a1, a2)
